Refactor: Modularize toolbar controls and sections, extract reusable utilities, and integrate streamlined toolbar structure with editor state syncing.
